# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Community Quiz API",
        "description": "Community-Based Quiz Management System API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Server"
        }
    ],
    "paths": {
...
        "/api/quizzes/{quiz}/start": {
            "post": {
                "tags": [
                    "Quiz Attempt"
                ],
                "summary": "Start a quiz attempt",
                "operationId": "quizAttemptStart",
                "parameters": [
                    {
                        "name": "quiz",
                        "in": "path",
                        "description": "Quiz ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "mode",
                                    "is_anonymous"
                                ],
                                "properties": {
                                    "mode": {
                                        "type": "string",
                                        "enum": [
                                            "practice",
                                            "scored"
                                        ]
                                    },
                                    "is_anonymous": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Attempt started"
                    },
                    "409": {
                        "description": "Active attempt exists"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/attempts/{attempt}/answer": {
            "post": {
                "tags": [
                    "Quiz Attempt"
                ],
                "summary": "Submit or update an answer (Save for later)",
                "operationId": "quizAttemptSubmitAnswer",
                "parameters": [
                    {
                        "name": "attempt",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "question_id"
                                ],
                                "properties": {
                                    "question_id": {
                                        "type": "integer"
                                    },
                                    "selected_option_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "answer_boolean": {
                                        "type": "boolean",
                                        "nullable": true
                                    },
                                    "answer_text": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Answer saved"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/attempts/{attempt}/submit": {
            "post": {
                "tags": [
                    "Quiz Attempt"
                ],
                "summary": "Submit a quiz attempt for scoring",
                "operationId": "quizAttemptSubmit",
                "parameters": [
                    {
                        "name": "attempt",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Attempt submitted"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/attempts/{attempt}": {
            "get": {
                "tags": [
                    "Quiz Attempt"
                ],
                "summary": "Get attempt details for resuming",
                "operationId": "quizAttemptShow",
                "parameters": [
                    {
                        "name": "attempt",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Attempt details"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/answers/{id}/grade": {
            "put": {
                "tags": [
                    "Quiz Attempt"
                ],
                "summary": "Grade a specific answer manually",
                "operationId": "quizAttemptGradeAnswerManual",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "score"
                                ],
                                "properties": {
                                    "score": {
                                        "type": "integer"
                                    },
                                    "feedback": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Answer graded"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/reviews/pending": {
            "get": {
                "tags": [
                    "Quiz Attempt"
                ],
                "summary": "List all quiz attempts pending manual review",
                "operationId": "quizAttemptPendingReviews",
                "responses": {
                    "200": {
                        "description": "Pending reviews"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter: Bearer {token}",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}